hmalloc: Add example.c for malloc/hmalloc testing
This patch adds a new example file, example.c, inside hmalloc/test
directory.  It is not for unittest but this can be useful when testing
hmctl. We can observe if proper amount of memory is allocated to the
intended target node with hmctl.

If user wants to set different memory policies for normal memory area
and hmalloc-ed area, then hmctl can be used after numactl as follows.

Usage 1)
  $ numactl -p 1 hmctl -m 2 ./example &
  [1] 49288
  255 MiB is allocated by malloc().
  511 MiB is allocated by hmalloc().
  Press enter to stop.

  $ numastat -c -p $!
  Per-node process memory usage (in MBs) for PID 49288 (example)
           Node 0 Node 1 Node 2 Node 3 Total
           ------ ------ ------ ------ -----
  Total         3    258    512      0   773

Usage 2)
  $ numactl -i 1,2 hmctl -p 3 ./example &
  [1] 49835
  255 MiB is allocated by malloc().
  511 MiB is allocated by hmalloc().
  Press enter to stop.

  $ numastat -c -p $!
  Per-node process memory usage (in MBs) for PID 49835 (example)
           Node 0 Node 1 Node 2 Node 3 Total
           ------ ------ ------ ------ -----
  Total         3    129    129    512   773

Signed-off-by: Honggyu Kim <honggyu.kim@sk.com>