hmalloc: Add -w/--weighted-interleave option to hmctl
This patch adds the -w/--weighted-interleave option to hmctl, which
sets a weighted interleave policy for hmalloc family allocations. It can
take multiple nodes for weighted interleave allocation, and memory will
be allocated in a weighted manner on the specified nodes.

The weights need to be set via its interface at the following location.

  /sys/kernel/mm/mempolicy/weighted_interleave/node*

Signed-off-by: Honggyu Kim <honggyu.kim@sk.com>
Signed-off-by: Yunjeong Mun <yunjeong.mun@sk.com>