tools: Calculate memory bandwidth ratio within the same package.
bwactl.py calculates memory bandwidth ratio for each numa nodes.
Weighted interleaving utilizes the memory bandwidth among memory
nodes in a directly linked package, not across processor interconnect
such as UPI.  So, for more accurate bandwidth ratio, it is calculated
only for the memory node within the same package.

Let's say there is a system having the following NUMA topology.
In this case, memory bandwidth ratio is separately calculated for
node 0,2 and node 1,3, which coexist in the same package.

    +-----------+     +-----------+
    |   node0   |     |   node1   |
    | +-------+ | UPI | +-------+ |
    | | CPU 0 |-+-----+-| CPU 1 | |
    | +-------+ |     | +-------+ |
    | | DRAM0 | |     | | DRAM1 | |
    | +---+---+ |     | +---+---+ |
    |     |     |     |     |     |
    | +---+---+ |     | +---+---+ |
    | | CXL 0 | |     | | CXL 1 | |
    | +---+---+ |     | +---+---+ |
    |   node2   |     |   node3   |
    +-----------+     +-----------+
      weighted          weighted
    interleaving      interleaving
       group 1           group 2

Signed-off-by: Yunjeong Mun <yunjeong.mun@sk.com>