tools: Rename gen_config.py to gen_migpol.py
The gen_config.py is used to generate damon migration policy, but the
name sounds vague so it'd be better to rename it to gen_migpol.py make
the meaning clear.

The meaning of migpol is migration policy so it clearly shows its
purpose because it generates a config file that contains DAMON
migration policy.

The output of gen_migpol.py can be directly given to damo start command.

Signed-off-by: Honggyu Kim <honggyu.kim@sk.com>