hmalloc: Refine nodemask handling in 'mbind' testcase
This patch modifies the method of setting the nodemask for memory
binding in the unittest of mbind. Previously, the nodemask was
hardcoded to '2', which was dependent on each test environment. This
patch changes the implementation to retrieve the nodemask using the
'numa_get_mems_allowed' function, making the tests compatiable with
different hardware environments.

Signed-off-by: Yunjeong Mun <yunjeong.mun@sk.com>
Signed-off-by: Honggyu Kim <honggyu.kim@sk.com>