tools: Add -s/--max-region-size option to gen_migpol.py
This patch adds the -s/--max-region-size option to gen_migpol.py that
defines the maximum size of a single region. When a region exceeds this
limit, it is split into multiple regions whose size does not exceed the
specified value. A separate monitoring context is created for each split
region, which results in multiple kdamond instances.

The example usage is as follows:

  $ sudo ./gen_migpol.py -m 0 2 -s 64G -o hmsdk.yaml

Suggested-by: Honggyu Kim <honggyu.kim@sk.com>
Signed-off-by: Yunjeong Mun <yunjeong.mun@sk.com>