hmalloc: Fix maxnode to avoid get_mempolicy failure
The current hmalloc_test fails as follows in some systems.

  $ ./hmalloc_test
      ...
  /home/honggyu/work/hmsdk/hmalloc/test/hmalloc_test.cpp:68: FAILED:
    CHECK( 0 == get_mempolicy(&hpolicy, &hnodemask, maxnode, addr, (1<<1)) )
  with expansion:
    0 == -1
      ...
  ===============================================================================
  test cases:   9 |   8 passed |  1 failed
  assertions: 186 | 171 passed | 15 failed

It fails when there are multiple channels of CXL memory attached and
detected as a single NUMA node.  For example, the failed configuration
is as follows.

  $ cat /sys/devices/system/node/online
  0-3

The onlined nodes are 4, but internally the possible value is higher as
follows.

  $ cat /sys/devices/system/node/possible
  0-11

This patch is to fix this problem by simply increasing the maxnode
number, but in the future, it is better to be compared with the possible
value.

Signed-off-by: Honggyu Kim <honggyu.kim@sk.com>