tools: Remove tools/gen_config.py
The tools/gen_config.py was used to generate a migration policy, but
replaced by tools/gen_migpol.py since hmsdk v3.0.

The gen_config.py was maintained as a symbolic link for backward
compatibility for a while, but it is now time to remove it to reduce
confusion with gen_migpol.py.

Signed-off-by: Honggyu Kim <honggyu.kim@sk.com>